vilaw_backend/requirements.txt
supabase-py
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-huggingface
langchain-community
langchain-pinecone
sentence-transformers
pydantic
pydantic-settings
pinecone-client
torch
transformers
python-docx
python-multipart    # Để nhận file upload từ API
pillow
psycopg2-binary